creates components for each section of the editor that requires complicated logic to remove blob and improve OO principles.
