+++ resources/assets/js/app.js
@@ -30,7 +30,7 @@ Vue.component('new-spot', require('./components/pages/home/NewSpot.vue').default
 Vue.component('filter-spots', require('./components/pages/home/FilterSpots.vue').default);
 Vue.component('admin-nav', require('./components/pages/admin/Nav.vue').default);
 Vue.component('admin-category-cards', require('./components/pages/admin/categories/CategoryCards.vue').default);
-Vue.component('admin-category-editor', require('./components/pages/admin/categories/CategoryEditor.vue').default);
+Vue.component('admin-category-editor', require('./components/pages/admin/categories/editor/CategoryEditor.vue').default);
 
 export const EventBus = new Vue();
